Significant change of architecture
Changed the architecture significantly.

Now each miner has their own copy of the blockchain.
All forms of validation have been moved under the miner class, as they are the actual ones doing the validation and checks.

The network has changed to be a better simulation of a network. It is essentially a place for miners to send blocks and accounts to send transactions.